deneme2

+++ javaKartvizit/src/main/java/com/example/dto/AuthResponse.java
@@ -35,21 +35,21 @@ public class AuthResponse {
     public void setUserInfo(UserInfo userInfo) { this.userInfo = userInfo; }
     
     public static class UserInfo {
-        private Long id;
+        private Integer id;
         private String username;
         private String email;
         
         public UserInfo() {}
         
-        public UserInfo(Long id, String username, String email) {
+        public UserInfo(Integer id, String username, String email) {
             this.id = id;
             this.username = username;
             this.email = email;
         }
         
         // Getters and Setters
-        public Long getId() { return id; }
-        public void setId(Long id) { this.id = id; }
+        public Integer getId() { return id; }
+        public void setId(Integer id) { this.id = id; }
         
         public String getUsername() { return username; }
         public void setUsername(String username) { this.username = username; }
